[UPDATE] server, get_tables, db
