+++ server/nginx-1.28.0/logs/access.log
@@ -6091,3 +6091,29 @@
 192.168.1.7 - - [16/Sep/2025:03:16:44 +0800] "GET /tbl_class_sched HTTP/1.1" 200 582 "-" "-"
 192.168.1.7 - - [16/Sep/2025:03:16:50 +0800] "GET /tbl_class_sched HTTP/1.1" 200 582 "-" "-"
 192.168.1.7 - - [16/Sep/2025:03:20:42 +0800] "GET /tbl_class_sched HTTP/1.1" 200 582 "-" "-"
+192.168.1.7 - - [16/Sep/2025:12:12:39 +0800] "GET /tbl_class_sched HTTP/1.1" 500 265 "-" "-"
+192.168.1.7 - - [16/Sep/2025:12:13:56 +0800] "GET /tbl_class_sched HTTP/1.1" 200 582 "-" "-"
+192.168.1.7 - - [16/Sep/2025:12:22:50 +0800] "GET /tbl_class_sched HTTP/1.1" 200 582 "-" "-"
+192.168.1.7 - - [16/Sep/2025:12:22:54 +0800] "GET /tbl_class_sched HTTP/1.1" 200 582 "-" "-"
+192.168.1.7 - - [16/Sep/2025:12:22:54 +0800] "GET /tbl_class_sched HTTP/1.1" 200 582 "-" "-"
+192.168.1.7 - - [16/Sep/2025:12:22:54 +0800] "GET /tbl_class_sched HTTP/1.1" 200 582 "-" "-"
+192.168.1.7 - - [16/Sep/2025:12:22:55 +0800] "GET /tbl_class_sched HTTP/1.1" 200 582 "-" "-"
+192.168.1.7 - - [16/Sep/2025:12:22:55 +0800] "GET /tbl_class_sched HTTP/1.1" 200 582 "-" "-"
+192.168.1.7 - - [16/Sep/2025:12:22:55 +0800] "GET /tbl_class_sched HTTP/1.1" 200 582 "-" "-"
+192.168.1.7 - - [16/Sep/2025:12:22:55 +0800] "GET /tbl_class_sched HTTP/1.1" 200 582 "-" "-"
+192.168.1.7 - - [16/Sep/2025:12:25:18 +0800] "GET /tbl_class_sched HTTP/1.1" 200 582 "-" "-"
+192.168.1.7 - - [16/Sep/2025:12:25:23 +0800] "GET /tbl_class_sched HTTP/1.1" 200 582 "-" "-"
+192.168.1.7 - - [16/Sep/2025:19:29:36 +0800] "GET /tbl_class_sched HTTP/1.1" 500 265 "-" "-"
+192.168.1.7 - - [16/Sep/2025:19:29:38 +0800] "GET /tbl_class_sched HTTP/1.1" 500 265 "-" "-"
+192.168.1.7 - - [16/Sep/2025:19:31:24 +0800] "GET /tbl_class_sched HTTP/1.1" 500 265 "-" "-"
+192.168.1.7 - - [16/Sep/2025:19:31:29 +0800] "GET /tbl_class_sched HTTP/1.1" 500 265 "-" "-"
+192.168.1.7 - - [16/Sep/2025:19:32:31 +0800] "GET /tbl_class_sched HTTP/1.1" 500 265 "-" "-"
+192.168.1.7 - - [16/Sep/2025:19:37:02 +0800] "GET /tbl_class_sched HTTP/1.1" 500 265 "-" "-"
+192.168.1.7 - - [16/Sep/2025:19:40:03 +0800] "GET /tbl_class_sched HTTP/1.1" 200 582 "-" "-"
+192.168.1.7 - - [16/Sep/2025:19:40:05 +0800] "GET /tbl_class_sched HTTP/1.1" 200 582 "-" "-"
+192.168.1.7 - - [16/Sep/2025:19:49:19 +0800] "GET /tbl_student_management HTTP/1.1" 404 207 "-" "-"
+192.168.1.7 - - [16/Sep/2025:19:49:22 +0800] "GET /tbl_class_sched HTTP/1.1" 200 582 "-" "-"
+192.168.1.7 - - [16/Sep/2025:19:49:28 +0800] "GET /tbl_student_management HTTP/1.1" 404 207 "-" "-"
+192.168.1.7 - - [16/Sep/2025:19:50:10 +0800] "GET /tbl_student_management HTTP/1.1" 404 207 "-" "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/140.0.0.0 Safari/537.36 Edg/140.0.0.0"
+192.168.1.7 - - [16/Sep/2025:19:50:11 +0800] "GET /favicon.ico HTTP/1.1" 404 207 "http://192.168.1.7:8080/tbl_student_management" "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/140.0.0.0 Safari/537.36 Edg/140.0.0.0"
+192.168.1.7 - - [16/Sep/2025:19:50:27 +0800] "GET /tbl_class_sched HTTP/1.1" 200 582 "-" "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/140.0.0.0 Safari/537.36 Edg/140.0.0.0"
