+++ server/nginx-1.28.0/logs/error.log
@@ -344,3 +344,11 @@
 2025/09/16 03:06:29 [notice] 15956#10268: signal process started
 2025/09/16 03:19:20 [notice] 16900#13380: signal process started
 2025/09/16 03:20:54 [notice] 5028#16484: signal process started
+2025/09/16 12:13:10 [notice] 1552#5536: signal process started
+2025/09/16 12:14:07 [notice] 18012#6620: signal process started
+2025/09/16 12:21:41 [notice] 14948#13236: signal process started
+2025/09/16 12:23:06 [notice] 17284#3340: signal process started
+2025/09/16 12:25:29 [notice] 13700#15812: signal process started
+2025/09/16 19:36:55 [notice] 17444#17544: signal process started
+2025/09/16 19:39:56 [notice] 19348#13476: signal process started
+2025/09/16 19:50:43 [notice] 18016#18572: signal process started
