+++ server/server_flask.py
@@ -7,6 +7,10 @@ app = Flask(__name__)
 def load_tbl_class_sched():
     return tbl_class_sched()
 
+@app.route('/tbl_student_management', methods=['GET'])
+def load_tbl_student_management():
+    return tbl_student_management()
+
 #Main
 if __name__ == '__main__':
     app.run(debug=True, host='0.0.0.0')
